feat(#40) : 공모전 팀 리스트 api 인터페이스 새엇ㅇ
